backend/requirements.txt
asgiref==3.10.0
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.14.0
dnspython==2.8.0
gunicorn==23.0.0
mongoengine==0.29.1
packaging==25.0
pymongo==4.15.3
python-dotenv==1.1.1
pytz==2025.2
sqlparse==0.5.3
